+++ xrpl-escrow-stdlib/src/ledger_objects/escrow.rs
@@ -1,7 +1,4 @@
-use xrpl_common_stdlib::core::ledger_objects::traits::LedgerObjectCommonFields;
-
-use crate::ledger_objects::traits::EscrowFields;
-
+use xrpl_common_stdlib::core::ledger_objects::traits::{EscrowFields, LedgerObjectCommonFields};
 #[derive(Debug, Clone, Copy, Eq, PartialEq)]
 pub struct Escrow {
     pub(crate) slot_num: i32,
@@ -18,5 +15,16 @@ impl EscrowFields for Escrow {}
 impl Escrow {
     pub fn new(slot_num: i32) -> Self {
         Self { slot_num }
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn test_new() {
+        let escrow = Escrow::new(42);
+        assert_eq!(escrow.slot_num, 42);
     }
 }
